Add the ability to provide graphName context to tasks
Using the graphName in a GsqlTask will apply all the commands under the
context of that graph. This is provided as a convenience so that you
don't need to provide the `USE GRAPH xxx` syntax within your gsql
scripts. This uses the `-g` command line parameter for the gsql client.